+++ .prettierrc
@@ -1,4 +1,4 @@
 {
-  "printWidth": 120,
+  "printWidth": 140,
   "singleQuote": false
 }
