+++ src/lib/extension/config.ts
