+++ src/lib/extension/index.ts
@@ -1,0 +1,2 @@
+export * from "./config";
+export * from "./hooks";
