+++ src/upcoming-events.tsx
@@ -1,11 +1,11 @@
+import { GoogleCalendarClient } from "@/lib/gcal/gcal-api-client";
 import { Action, ActionPanel, Color, Icon, List, Toast, showToast } from "@raycast/api";
 import { getAccessToken, withAccessToken } from "@raycast/utils";
-import { format, intervalToDuration, formatDistance, formatDuration } from "date-fns";
-import { useConfig, useUpcomingEvents } from "./lib/hooks";
-import { RaycastGoogleOAuthService } from "./lib/raycast/google-oauth-service";
-import { GoogleCalendarEvent } from "./lib/models";
+import { format, formatDuration, intervalToDuration } from "date-fns";
 import { useMemo, useState } from "react";
-import { GoogleCalendarClient } from "./lib/api-client";
+import { useConfig, useUpcomingEvents } from "./lib/extension/hooks";
+import { GoogleCalendarEvent } from "./lib/gcal/models";
+import { RaycastGoogleOAuthService } from "./lib/gcal/raycast-google-auth-service";
 
 function Command() {
   const { data: events, isLoading: isEventsLoading, revalidate: refreshEvents } = useUpcomingEvents();
